Development.

+++ Database/classes/MDatabase.php
@@ -210,7 +210,7 @@ class MDatabase extends E\Module
 		if ($date === null)
 			return null;
 
-		return strtotime($date . ' UTC');
+		return (int)strtotime($date . ' UTC');
 	}
 
 	// public function unescapeTimeDateMillis($date)
@@ -224,9 +224,9 @@ class MDatabase extends E\Module
 	public function unescapeTime_DateTime($date_time)
 	{
 		if ($date_time === null)
-			return null;
+            return null;
 
-		return strtotime($date_time . ' UTC');
+		return (int)strtotime($date_time . ' UTC');
 	}
 
 	public function unescapeInt($value)
